modularización de interfaces
