+++ src/components/campaign/CampaignCard.tsx
@@ -1,5 +1,5 @@
 // src/components/campaigns/CampaignCard.tsx
-import { ICampaign } from "@/interfaces";
+import { ICampaign } from "@/interfaces/campaign.interface";
 import React from "react";
