+++ src/interfaces/AuthInterfaces/login.interfaces.ts
@@ -1,0 +1,18 @@
+export interface ILoginProps {
+  email: string;
+  password: string;
+}
+
+export interface ILoginErrors {
+  email?: string;
+  password?: string;
+}
+
+export interface ILoginResponse {
+  token: string;
+  user: {
+    id: string;
+    name: string;
+    email: string;
+  };
+}
