+++ src/interfaces/AuthInterfaces/register.interfaces.ts
@@ -1,0 +1,26 @@
+export interface IRegisterValues {
+  name: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+}
+
+export interface IRegisterErrors {
+  name?: string;
+  email?: string;
+  password?: string;
+  confirmPassword?: string;
+}
+
+export interface IRegisterProps {
+  name: string;
+  email: string;
+  password: string;
+}
+
+export interface IRegisterResponse {
+  id: string;
+  name: string;
+  email: string;
+  token?: string;
+}
