+++ src/interfaces/AuthInterfaces/user.interface.ts
@@ -1,0 +1,9 @@
+export interface IUserSession {
+    token: string,
+    user: {
+        id: number,
+        email: string,
+        name: string,
+        donations: []
+    }
+}
