+++ src/interfaces/campaign.interface.ts
@@ -1,0 +1,10 @@
+export interface ICampaign {
+  title: string;
+  description: string;
+  startDate: string;
+  id: number;
+  endDate: string;
+  isActive: boolean;
+}
+
+
